+++ packages/spacecat-shared-cloud-manager-client/CHANGELOG.md
@@ -1,3 +1,43 @@
+## [@adobe/spacecat-shared-cloud-manager-client-v1.2.2](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-cloud-manager-client-v1.2.1...@adobe/spacecat-shared-cloud-manager-client-v1.2.2) (2026-05-14)
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#1533](https://github.com/adobe/spacecat-shared/issues/1533)) ([0a3e2ab](https://github.com/adobe/spacecat-shared/commit/0a3e2abbbc5f58b5320518f7d596d4cef6271fa0))
+
+## [@adobe/spacecat-shared-cloud-manager-client-v1.2.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-cloud-manager-client-v1.2.0...@adobe/spacecat-shared-cloud-manager-client-v1.2.1) (2026-05-05)
+
+### Bug Fixes
+
+* **cloud-manager-client:** re-publish after revert of submodule cloning ([#1583](https://github.com/adobe/spacecat-shared/issues/1583)) ([c22b933](https://github.com/adobe/spacecat-shared/commit/c22b93333391ef869d8194225be59a5e36c200ba))
+
+### Reverts
+
+* "feat: CM git clone also clones submodules ([#1583](https://github.com/adobe/spacecat-shared/issues/1583))" ([#1587](https://github.com/adobe/spacecat-shared/issues/1587)) ([2d9e34e](https://github.com/adobe/spacecat-shared/commit/2d9e34e68a0af6612891c61ec37e4304b7e9d912))
+
+## [@adobe/spacecat-shared-cloud-manager-client-v1.2.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-cloud-manager-client-v1.1.5...@adobe/spacecat-shared-cloud-manager-client-v1.2.0) (2026-05-05)
+
+### Features
+
+* CM git clone also clones submodules ([#1583](https://github.com/adobe/spacecat-shared/issues/1583)) ([a1e67c3](https://github.com/adobe/spacecat-shared/commit/a1e67c36f9fb01dd65b501a2a1791f18a7432465))
+
+## [@adobe/spacecat-shared-cloud-manager-client-v1.1.5](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-cloud-manager-client-v1.1.4...@adobe/spacecat-shared-cloud-manager-client-v1.1.5) (2026-04-20)
+
+### Reverts
+
+* Revert "fix: clone functionality to include submodules ([#1413](https://github.com/adobe/spacecat-shared/issues/1413))" ([#1549](https://github.com/adobe/spacecat-shared/issues/1549)) ([f4594d4](https://github.com/adobe/spacecat-shared/commit/f4594d4a22ddecf78fef71510943f63731203ab7))
+
+## [@adobe/spacecat-shared-cloud-manager-client-v1.1.4](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-cloud-manager-client-v1.1.3...@adobe/spacecat-shared-cloud-manager-client-v1.1.4) (2026-04-18)
+
+### Bug Fixes
+
+* clone functionality to include submodules ([#1413](https://github.com/adobe/spacecat-shared/issues/1413)) ([a5c0016](https://github.com/adobe/spacecat-shared/commit/a5c0016987f6ebb8235493f877f1a681326192ec))
+
+## [@adobe/spacecat-shared-cloud-manager-client-v1.1.3](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-cloud-manager-client-v1.1.2...@adobe/spacecat-shared-cloud-manager-client-v1.1.3) (2026-04-15)
+
+### Bug Fixes
+
+* **cm-client:** preserve broken symlinks when zipping repositories ([#1468](https://github.com/adobe/spacecat-shared/issues/1468)) ([943662a](https://github.com/adobe/spacecat-shared/commit/943662ac5d3765ee2a289355b9bf1e23c6c88230))
+
 ## [@adobe/spacecat-shared-cloud-manager-client-v1.1.2](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-cloud-manager-client-v1.1.1...@adobe/spacecat-shared-cloud-manager-client-v1.1.2) (2026-04-06)
 
 ### Bug Fixes